reorganized utils/map-utils
