+++ bikespace_frontend/src/utils/index.ts
@@ -1,2 +1,1 @@
 export * from './track-umami-event';
-export * from './map-utils/mapUtils';
