+++ bikespace_frontend/src/utils/map-utils/geocoder-search/geocoder-search.module.scss
@@ -1,0 +1,1 @@
+@import '@/styles/variables.module.scss';
