+++ bikespace_frontend/src/utils/map-utils/index.ts
@@ -1,0 +1,2 @@
+export * from './mapUtils';
+export * from './geocoder-search/GeocoderSearch';
